requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
torch
transformers
sentence-transformers
